Fix flatMap overloading issues
Overloading resolution is mostly done on the first parameter list, but for an
extension method, the first parameter list is just the receiver, so overloading
resolution failed to disambiguate between the two extension methods named
`flatMap`.

It turns out that `flatMap` can be an instance method on `Query` and doesn't
need to be an extension method since it's defined for all Query, so just moving
it in `Query` fixes the overloading issues. Most of the over extension methods
should be moved too.