Fix bug in `multiFix` masked by cast
The type of `val unions` used to be `List[Unit]` because the last line in the
map was `idx += 1` which returns Union. This did not lead to a compile-time
error because of later casting, but it lead to a ClassCastException when
running the RecursiveTwoMultiTest test.

This commit rewrites `unions` to use a zip instead of manually incrementing an
index, we also make use of type bindings to locally reduce the amount of casting
needed.

The test doesn't crash anymore but still fails as the expected SQL string does
not match the generated one.