Rework `multiFix` to be able to infer its type parameters
Previously we had a signature that started with:

    def multiFix[T <: Tuple](bases: Tuple.Map[T, Query])(...)

But if I pase `(Query[A], Query[B], ...)` to this method, the compiler is unable
to infer `T` (it doesn't know how to inverse `Tuple.Map`). To work around this
we can instead define:

    def multiFix[QT <: Tuple](bases: QT)(using Tuple.Union[QT] <:< Query[?])(...)

and use `Tuple.InverseMap[QT, Query]` to get back what we previously called `T`.
This works OK but there is an additional subtlety to consider, the full
signature of multiFix used to be:

   def multiFix[T <: Tuple](bases: Tuple.Map[T, Query])(fns: Tuple.Map[T, QueryRef] => Tuple.Map[T, Query]): Tuple.Map[T, Query] =

And so one might expect that the new full signature should be:

   type Elems[QT <: Tuple] = Tuple.InverseMap[QT, Query]
   type ToQueryRef[QT <: Tuple] = Tuple.Map[Elems[QT], QueryRef]

   def multiFix[QT <: Tuple](bases: QT)(using Tuple.Union[QT] <:< Query[?])(fns: ToQueryRef[QT] => QT): QT =

But this does not work, in particular in `RecursiveTwoMultiTest` we have:

    val pathBase: Table[Edge] = ...
    val pathToABase: Table[Edge] = ...
    val (pathResult, pathToAResult) = multiFix(pathBase, pathToABase)((path, pathToA) =>
      // ...
      someValue : (Query[Edge], Query[Edge])
    )

The issue is that the first argument we pass to `multiFix` has type
`(Table[Edge], Table[Edge])`, so this is what type inference picks up for `QT`,
but we really want the lambda to be able to return an arbitrary tuple of
`(Query[Edge], Query[Edge])`, so we need an extra seemingly unnecessary mapping
to obtain the final full signature:

   type ToQuery[QT <: Tuple] = Tuple.Map[Elems[QT], Query]
   def multiFix[QT <: Tuple](bases: QT)(using Tuple.Union[QT] <:< Query[?])(fns: ToQueryRef[QT] => ToQuery[QT]): ToQuery[QT] =